requirements.txt
asgiref==3.6.0
click==8.1.3
dj-database-url==1.2.0
dj-email-url==1.0.6
dj-rest-auth==3.0.0
Django==4.1.7
django-cache-url==3.4.4
django-extensions==3.2.1
djangorestframework==3.14.0
environs==9.5.0
flake8==6.0.0
flake8-html==0.4.3
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.0
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
